// src/app/model/periods-bet.ts
export interface graphItem {
	pone: number;
	psix: number;
	ptwo: number;
	pfive: number;
	pthree: number;
	pfour: number;
}

export interface goalsItem {
	pone: number;
	psix: number;
	ptwo: number;
	pfive: number;
	pthree: number;
	pfour: number;
}

export interface PeriodsBet {
	graph: graphItem;
	goals: goalsItem;
  goalsAgg: goalsItem;
}